# [network][connmgr] reset both address and backoff on update