--- a/network/src/validator_network/network_builder.rs
+++ b/network/src/validator_network/network_builder.rs
@@ -63,7 +63,7 @@ pub const MAX_CONCURRENT_INBOUND_RPCS: u32 = 100;
 pub const PING_FAILURES_TOLERATED: u64 = 10;
 pub const MAX_CONCURRENT_NETWORK_REQS: usize = 100;
 pub const MAX_CONCURRENT_NETWORK_NOTIFS: usize = 100;
-pub const MAX_CONNECTION_DELAY_MS: u64 = 1 * 60 * 1000; // 1 minute
+pub const MAX_CONNECTION_DELAY_MS: u64 = 60_000; /* 1 minute */
 
 #[derive(Debug)]
 pub enum AuthenticationMode {