@andll 2020-06-11
Question - do you want to get any data points to actually confirm how this change affected recovery?

As a cluster test advertising - one option for this, for example 10% down experiment and compare graphs between master and this diff to see if recovery time improves?
You could do this by those two commands:
```
# Reference
./scripts/cti --master --run bench -- --percent-nodes-down 10
# This PR
./scripts/cti --pr <PR> --run bench -- --percent-nodes-down 10
```
They print link to dashboard in the end, so in theory those links can be used to see a difference