@phlip9 2020-06-11
looks like the cluster initially connects much more quickly with this fix 

current master:

![Screen Shot 2020-06-11 at 13 25 57](https://user-images.githubusercontent.com/918989/84435800-41ebf800-abe7-11ea-9683-81f0c5a754dc.png)

with pr:

![Screen Shot 2020-06-11 at 13 27 34](https://user-images.githubusercontent.com/918989/84435856-57f9b880-abe7-11ea-9921-2eb87a3e4273.png)