@andll 2020-06-11
Thats awesome!
First picture(master) is exactly problem that we seen today with @sherry-x  in premainnet drill