@phlip9 2020-06-11
forgot the top level metrics : p

```
before: 10% down : 1285 TPS, 3160 ms latency, 7550 ms p99 latency, no expired txns
after:  10% down : 1333 TPS, 3050 ms latency, 4350 ms p99 latency, no expired txns
```